requirements.txt
langchain-google-genai
langchain
langchain-community
python-dotenv
